… src/pixie/images.nim
… src/pixie/paths.nim
… tests/benchmark_jpeg.nim
… tests/test_png.nim