png.nim

before:
```
pixie decode ....................... 4.125 ms      4.167 ms    ±0.031  x1000
```
after:
```
pixie decode ....................... 3.690 ms      3.722 ms    ±0.026  x1000
```

internal.nim

before:
```
fill_rgba .......................... 0.289 ms      0.310 ms    ±0.015  x1000
```

after:
```
fill_rgba .......................... 0.270 ms      0.279 ms    ±0.011  x1000
```